bulk change: changes related to JSDoc issue
We decided to use a JSDuck as a tool for generating API docs.
During its work JSDuck parses JS code. JSDocs are generated from
raw files. They contain patterns that is being substituted with
values during build process. This requres raw files to be a valid
JS scripts. Due to this we changed start/end tokens for all
substitutions.