Added one more color module
It turned out that there are some quantity of javascript libs that work with colors. Due to this we need to start using one of that libs and drop as much as possible of our code.

For now best way to go is to use static file of one libs.

For the reference see Issue #65