feat(build) do not include functions for HTML-based script into JS-based
Running long-standing loops in HTML-based scripts may cause a UI freezes
and other side effects. To fix that we need to call `Host.CanContinue()`
method every 5 seconds (not sure about interval duration). However that
call is useless in JS-based scripts. In this commit we add support of
custom marker that can be replaced with text/code that is correspondent
to current type of build script

See issue #9 for details